liking and following enabled
Functionality for liking tweets and following users has been implemented. Still need to connect functionality to the database.
#24

Some other things that were changed:
Removed tweet image from the design
Changed the style of the like button
Added textview for a like counter next to the like button
Made it so you can click on a user's profile from their tweet and be sent to their profile page
Can also click on your own profile picture on the timeline page and be brought to your own profile
changed error messages on composeTweetActivity to toast messages